ai/db-hackathon2023/chat-ui-react/examples/stocks_recommendation_api/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-storage==1.28.1
Flask-Cors==4.0.0
